% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bar_data}
\alias{bar_data}
\title{Function to generate aggregated data per time}
\arguments{
\item{choice_of_aggregation}{\code{character} How you would like to aggregate the listened tracks}
}
\description{
Function to generate aggregated data per time
}
