% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barplots}
\alias{barplots}
\title{Get a simple barplot by week}
\arguments{
\item{choice_of_aggregation}{\code{character} How you would like to aggregate the listened tracks}
}
\description{
Get a simple barplot by week
}
